refactor: relocate page components into the appropriate pages directory
