+++ src/App.tsx
@@ -29,7 +29,7 @@ import './App.css'
 import PointShop from './pages/point-shop'
 import ProductDetail from './pages/point-shop/products/[pointProductId]/detail'
 import PostDetail from './pages/posts/[id]'
-import EnrollAddress from './components/shop-screen/EnrollAddress'
+import EnrollAddress from './pages/point-shop/products/[pointProductId]/enrollAddress'
 import Signup from './pages/signup'
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools'
 import Posts from './pages/posts'
