+++ src/pages/point-shop/products/[pointProductId]/enrollAddress.tsx
@@ -1,9 +1,9 @@
 import { ChevronLeft } from 'lucide-react'
-import AddressInput, { AddressState } from '../common/form/AddressInput'
+import AddressInput, { AddressState } from '../../../../components/common/form/AddressInput'
 import { useCallback, useState } from 'react'
-import { Input } from '../ui/input'
-import InputLabel from '../common/form/InputLabel'
-import BottomNavigation from '../common/BottomNav'
+import { Input } from '../../../../components/ui/input'
+import InputLabel from '../../../../components/common/form/InputLabel'
+import BottomNavigation from '../../../../components/common/BottomNav'
 import { useLocation, useNavigate, useSearchParams } from 'react-router-dom'
 import { addressApi } from '@/api/address'
 import { ServerPostAddress } from '@/types/addresses'
